refactor(core): fold searchQueryLength into search_debouncer
Move the filter-length extractor and its tests into search_debouncer.dart /
search_debouncer_test.dart (alongside the debounce policy that consumes the
length) and drop the standalone search_query_length files.
