+++ packages/stream_chat_flutter_core/lib/src/stream_member_list_controller.dart
@@ -7,7 +7,6 @@ import 'package:stream_chat/stream_chat.dart' hide Success;
 import 'package:stream_chat_flutter_core/src/paged_value_notifier.dart';
 import 'package:stream_chat_flutter_core/src/search_debounce_mixin.dart';
 import 'package:stream_chat_flutter_core/src/search_debouncer.dart';
-import 'package:stream_chat_flutter_core/src/search_query_length.dart';
 
 /// The default channel page limit to load.
 const defaultMemberPagedLimit = 10;
